WIP downgrading to go 1.18 since that is what rclone is stuck on

+++ deafult_client_stress_test.go
@@ -24,7 +24,7 @@ func TestDefaultClient_Photos_Stress(t *testing.T) {
 	const stressTestPhotoCount = 120
 	const maxConcurrentRequests = 20
 
-	assert.Greater(t, stressTestPhotoCount, photoPageSize)
+	assert.Greater(t, uint64(stressTestPhotoCount), photoPageSize)
 
 	type testData struct {
 		containerType types.ContainerType
